+++ Sources/Controllers/ALKConversationListViewController.swift
@@ -50,6 +50,7 @@ open class ALKConversationListViewController: ALKBaseViewController, Localizable
             style: .plain,
             target: self, action: #selector(compose)
         )
+        barButton.accessibilityIdentifier = "NewChatButton"
         return barButton
     }()
 
@@ -71,6 +72,7 @@ open class ALKConversationListViewController: ALKBaseViewController, Localizable
         localizedStringFileName = configuration.localizedStringFileName
     }
 
+    @available(*, unavailable)
     public required init?(coder _: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
@@ -82,7 +84,7 @@ open class ALKConversationListViewController: ALKBaseViewController, Localizable
         conversationListTableViewController.remove()
     }
 
-    open override func addObserver() {
+    override open func addObserver() {
         NotificationCenter.default.addObserver(forName: NSNotification.Name(rawValue: "newMessageNotification"), object: nil, queue: nil, using: { [weak self] notification in
             guard let weakSelf = self else { return }
             let msgArray = notification.object as? [ALMessage]
@@ -176,7 +178,7 @@ open class ALKConversationListViewController: ALKBaseViewController, Localizable
         })
     }
 
-    open override func removeObserver() {
+    override open func removeObserver() {
         NotificationCenter.default.removeObserver(self, name: NSNotification.Name(rawValue: "pushNotification"), object: nil)
         NotificationCenter.default.removeObserver(self, name: NSNotification.Name(rawValue: "newMessageNotification"), object: nil)
         NotificationCenter.default.removeObserver(self, name: NSNotification.Name(rawValue: "reloadTable"), object: nil)
@@ -185,7 +187,7 @@ open class ALKConversationListViewController: ALKBaseViewController, Localizable
         NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardDidHideNotification, object: nil)
     }
 
-    open override func viewWillAppear(_ animated: Bool) {
+    override open func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         activityIndicator.center = CGPoint(x: view.bounds.size.width / 2, y: view.bounds.size.height / 2)
         activityIndicator.color = UIColor.gray
@@ -194,7 +196,7 @@ open class ALKConversationListViewController: ALKBaseViewController, Localizable
         viewModel.prepareController(dbService: dbService)
     }
 
-    open override func viewDidLoad() {
+    override open func viewDidLoad() {
         super.viewDidLoad()
         alMqttConversationService = ALMQTTConversationService.sharedInstance()
         alMqttConversationService.mqttConversationDelegate = self
@@ -206,7 +208,7 @@ open class ALKConversationListViewController: ALKBaseViewController, Localizable
         extendedLayoutIncludesOpaqueBars = true
     }
 
-    open override func viewDidAppear(_: Bool) {
+    override open func viewDidAppear(_: Bool) {
         print("contact id: ", contactId as Any)
         if contactId != nil || channelKey != nil || conversationId != nil {
             print("contact id present")
@@ -320,7 +322,8 @@ open class ALKConversationListViewController: ALKBaseViewController, Localizable
         if let viewController = conversationViewController,
             viewController.viewModel != nil,
             viewController.viewModel.contactId == message.contactId,
-            viewController.viewModel.channelKey == message.groupId {
+            viewController.viewModel.channelKey == message.groupId
+        {
             print("Contact id matched1")
             viewController.viewModel.addMessagesToList([message])
         }
@@ -335,7 +338,7 @@ open class ALKConversationListViewController: ALKBaseViewController, Localizable
         }
     }
 
-    open override func showAccountSuspensionView() {
+    override open func showAccountSuspensionView() {
         let accountVC = ALKAccountSuspensionController()
         accountVC.closePressed = { [weak self] in
             self?.dismiss(animated: true, completion: nil)
@@ -450,7 +453,8 @@ extension ALKConversationListViewController: ALMQTTConversationDelegate {
         if let vm = viewController?.viewModel, vm.contactId != nil || vm.channelKey != nil,
             let visibleController = navigationController?.visibleViewController,
             visibleController.isKind(of: ALKConversationViewController.self),
-            isNewMessageForActiveThread(alMessage: alMessage, vm: vm) {
+            isNewMessageForActiveThread(alMessage: alMessage, vm: vm)
+        {
             viewModel.syncCall(viewController: viewController, message: message, isChatOpen: true)
 
         } else if !isMessageSentByLoggedInUser(alMessage: alMessage) {
@@ -464,7 +468,8 @@ extension ALKConversationListViewController: ALMQTTConversationDelegate {
             }
         }
         if let visibleController = navigationController?.visibleViewController,
-            visibleController.isKind(of: ALKConversationListViewController.self) {
+            visibleController.isKind(of: ALKConversationListViewController.self)
+        {
             sync(message: alMessage)
         }
     }