+++ Sources/Controllers/ALKConversationViewController+NavigationBar.swift
@@ -5,8 +5,8 @@
 //  Created by Mukesh on 22/06/20.
 //
 
-import Foundation
 import Applozic
+import Foundation
 
 extension ALKConversationViewController: NavigationBarCallbacks {
     open func titleTapped() {
@@ -35,8 +35,9 @@ extension ALKConversationViewController: NavigationBarCallbacks {
         guard
             viewModel.channelKey == nil,
             viewModel.conversationProxy == nil,
-            let contactId = viewModel.contactId else {
-                return nil
+            let contactId = viewModel.contactId
+        else {
+            return nil
         }
         return ALContactService().loadContact(byKey: "userId", value: contactId)
     }