+++ Sources/Controllers/ALKNewChatViewController.swift
@@ -42,6 +42,7 @@ public final class ALKNewChatViewController: ALKBaseViewController, Localizable 
         setupView()
     }
 
+    @available(*, unavailable)
     required init?(coder _: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
@@ -50,7 +51,7 @@ public final class ALKNewChatViewController: ALKBaseViewController, Localizable 
         super.init(configuration: configuration)
     }
 
-    public override func viewDidLoad() {
+    override public func viewDidLoad() {
         super.viewDidLoad()
         searchBar.delegate = self
         ALUserDefaultsHandler.setContactServerCallIsDone(false)
@@ -60,7 +61,7 @@ public final class ALKNewChatViewController: ALKBaseViewController, Localizable 
         }
     }
 
-    public override func viewWillAppear(_ animated: Bool) {
+    override public func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         edgesForExtendedLayout = []
         activityIndicator.center = CGPoint(x: view.bounds.size.width / 2, y: view.bounds.size.height / 2)