+++ Sources/Views/ALKFormPasswordItemCell.swift
@@ -43,7 +43,8 @@ class ALKFormPasswordItemCell: UITableViewCell {
         addConstraints()
     }
 
-    required init?(coder: NSCoder) {
+    @available(*, unavailable)
+    required init?(coder _: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }