+++ Sources/Views/ALKFormTextItemCell.swift
@@ -42,7 +42,8 @@ class ALKFormTextItemCell: UITableViewCell {
         addConstraints()
     }
 
-    required init?(coder: NSCoder) {
+    @available(*, unavailable)
+    required init?(coder _: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
 
@@ -87,7 +88,8 @@ class ALKFormItemHeaderView: UITableViewHeaderFooterView {
         addConstraints()
     }
 
-    required init?(coder: NSCoder) {
+    @available(*, unavailable)
+    required init?(coder _: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }