+++ Sources/Views/ALKPhotoCell.swift
@@ -14,7 +14,8 @@ import UIKit
 // MARK: - ALKPhotoCell
 
 class ALKPhotoCell: ALKChatBaseCell<ALKMessageViewModel>,
-    ALKReplyMenuItemProtocol, ALKReportMessageMenuItemProtocol {
+    ALKReplyMenuItemProtocol, ALKReportMessageMenuItemProtocol
+{
     var photoView: UIImageView = {
         let mv = UIImageView()
         mv.backgroundColor = .clear
@@ -380,7 +381,7 @@ class ALKPhotoCell: ALKChatBaseCell<ALKMessageViewModel>,
         dbMessage.fileMetaInfo.thumbnailFilePath = filePath
 
         let dbHandler = ALDBHandler.sharedInstance()
-        let error =  dbHandler?.saveContext()
+        let error = dbHandler?.saveContext()
         if error != nil {
             print("Not saved due to error \(String(describing: error))")
         }